rename session to persistence
